Made the keyboard listener suppress key events
Now, pressing the arrow keys, enter, or 'c' while running tetris.py
will not interfere with anything in the terminal window or any
other windows. The downside is that this makes the rest of the
computer basically unusable for the user, so I'll try to find
another method for monitoring key presses that depends on
window context or something like that.

To exit out of the program, press Escape.