Added basic functionality for the right analog stick
The user can now alternate what type of opponent gets targeted
by pressing 'A' on the keyboard. Each press of the A key makes the
emulated right analog stick alternate between all directions
(as well as the center), so all types of opponents can be
targeted in Tetris 99.

This code is very rough right now and definitely needs improvement
/ error handling, but it adds a lot of functionality to the
project.