Changed the right joystick controls on the keyboard
Now, the WASD keys are used to make the Switch's right control
stick point in a specific direction. When one of these keys is
pressed, the program will hold the simulated joystick for a
quarter of a second to make sure that the switch recognizes the
new direction and changes who is being targeted in Tetris 99.