Improved analog stick controls
Instead of making the WASD keys push the analog stick in a specific
direction for a constant amount of time, the script now puts the
stick in a specific direction for the exact amount of time each
direction key is pressed down (just like the button controls). This
makes the analog stick controls feel much more responsive as a
result.